tools/dialogue_server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.42.0
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
